+++ ReleaseNotes.md
@@ -1,3 +1,17 @@
+# 0.2.0
+
+- Update to pallet 0.8.0-RC.9
+
+- Tags on new instances reflected in Nodes
+
+- Tag instance names
+
+- Add ami-info function to parse ami descriptions
+
+- Remove AMI description parsing for Node info
+
+- Implement ComputeServiceProperties
+
 # 0.1.1
 
 - Change describe-key-pairs to use :key-names